Deprecate reading depth when it was changed
Typically, State#depth is the same value at the start of `generate` that
it is upon returning from it. However when a NestingError is raised, it
is not reset to its initial value.

This behavior is not that useful given the value will always be equal to
`max_nesting` in those cases, and unchanged otherwise. Keeping this
behavior prevents us from removing duplication and removing
`State#generate_new`.